fix: 815 - docs menu rerenders
