+++ src/app/docs/[lang]/[...slug]/page.tsx
@@ -1,7 +1,7 @@
 import { Metadata } from 'next';
 import { notFound } from 'next/navigation';
 
-import { DocsLayout } from '../../components/docs-layout/docs-layout';
+import { DocsContent } from '../../components/docs-content/docs-content';
 import { TITLE_POSTFIX } from '../../constants';
 import { Menu } from '../../types';
 import { fetchMarkdownContent } from '../../utils/fetchMarkdownContent';
@@ -90,19 +90,14 @@ export async function generateStaticParams(): Promise<RouteParams[]> {
   return allSlugs;
 }
 
-export default async function DocPage({
-  params,
-}: {
-  params: Promise<RouteParams>;
-}) {
+export default async function DocPage({ params }: { params: Promise<RouteParams> }) {
   const { lang, slug } = await params;
 
   try {
     const markdownContent = await fetchMarkdownContent(lang, slug);
-    const docsMenu = await fetchMenu(lang);
 
-    return <DocsLayout menu={docsMenu} markdownContent={markdownContent} lang={lang} />;
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    return <DocsContent markdownContent={markdownContent} lang={lang} />;
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
   } catch (error) {
     notFound();
   }
